Rename SidebarAdmin to sidebar-admin, update imports and CSS classes accordingly; refactor NavBar and PaymentSection components
