+++ src/app/admin/manage-oven-detail/page.tsx
@@ -1,12 +1,12 @@
 "use client";
-import IngredientDetail from "@/components/Admin/manage-ingredient-detail";
-import Oven from "@/components/Admin/manage-oven-detail";
+import IngredientDetail from "@/components/admin/manage-ingredient-detail";
+import Oven from "@/components/admin/manage-oven-detail";
 import Topping, {
   ToppingDetail,
-} from "@/components/Admin/manage-topping-detail";
+} from "@/components/admin/manage-topping-detail";
 import Button from "@/components/button/button";
 import SearchBar from "@/components/search-bar/search-bar";
-import SidebarAdmin from "@/components/SidebarAdmin/sidebarAdmin";
+import SidebarAdmin from "@/components/sidebar-admin/sidebarAdmin";
 
 const ManageOvenDetail = () => {
   const pseudoOvenData = {
