+++ src/app/product/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import Footer from "@/components/Footer/footer";
+import Footer from "@/components/footer/footer";
 import Header from "@/components/header/header";
 import NavBar from "@/components/header/nav-bar";
 import DetailProduct, {
