+++ src/components/header/header.css
@@ -1,4 +1,4 @@
-.menu {
+.nav-menu {
   /* font-family: Poppins, sans-serif; */
   margin-right: 2px;
   overflow: hidden;
